refactor: scripts and engineer use same commands all commands were copy-pasted from the scripts cli, which caused the multiple cli's to be unmaintainable.
