+++ packages/scripts/src/index.ts
@@ -15,3 +15,4 @@ export * from './node-environments-manager';
 export * from './types';
 export * from './utils';
 export * from './ws-environment';
+export * from './cli-commands';
